Actualizacion Sabado

+++ BecarioAPI/Controllers/UsuarioController.cs
@@ -1,24 +1,23 @@
 ﻿using BecarioAPI.Models.Interfaces;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BecarioAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class UsuarioController : Controller
+    public class UsuarioController : ControllerBase
     {
-        //Inyeccion de dependencias
-        #region Dependencias
+        //Inyección de dependencias
         private readonly IUsuarioRepository _usuarioRepository;
-        private readonly ILogger<UsuarioController> _logger;
+        private readonly ILogger<UsuarioController> _logger;   
         public UsuarioController(IUsuarioRepository usuarioRepository, ILogger<UsuarioController> logger)
         {
             _usuarioRepository = usuarioRepository;
             _logger = logger;
         }
-        #endregion
 
-        [HttpPost]
+        [HttpPost("Login")]
         public ActionResult<bool> Login(string email, string password)
         {
             try
@@ -35,28 +34,45 @@ namespace BecarioAPI.Controllers
         }
 
         [HttpPost("AgregarUsuario")]
-        public void AddUsuario(Becas.Shared.Usuario usuario)
+        public ActionResult AddUsuario([FromBody] Becas.Shared.Usuario usuario)
         {
             try
             {
+                //validar que el usuario no exista
+                if (_usuarioRepository.GetUsuario(usuario.Email) != null)
+                {
+                    throw new Exception("El usuario ya existe");
+                }
                 _usuarioRepository.AddUsuario(usuario);
+                return Ok("Usuario Agragado Exitosamente");
+
             }
             catch (Exception ex)
             {
                 _logger.LogError($"Error en {nameof(AddUsuario)}: " + ex.Message);
+                return BadRequest(ex.Message);
             }
+           
         }
 
-        [HttpDelete("EliminarUsuario")]
-        public void DeleteUsuario(string email)
+        [HttpDelete("EliminarUsuario/{email}")]
+        public ActionResult DeleteUsuario(string email)
         {
             try
             {
+                //validar que el usuario exista
+                var usuario = _usuarioRepository.GetUsuario(email);
+                if(usuario == null)
+                {
+                    throw new Exception("El usuario no existe");
+                }
                 _usuarioRepository.DeleteUsuario(email);
+                return Ok("Usuario Eliminado Exitosamente");
             }
             catch (Exception ex)
             {
                 _logger.LogError($"Error en {nameof(DeleteUsuario)}: " + ex.Message);
+                return BadRequest(ex.Message);
             }
         }
 
@@ -66,8 +82,13 @@ namespace BecarioAPI.Controllers
         {
             try
             {
+                //validar que el usuario exista
                 var usuario = _usuarioRepository.GetUsuario(email);
-                return usuario;
+                if(usuario == null)
+                {
+                    throw new Exception("El usuario no existe");
+                }
+                return Ok(usuario);
             }
             catch (Exception ex)
             {
@@ -81,8 +102,13 @@ namespace BecarioAPI.Controllers
         {
             try
             {
+               //validar que existan usuarios
                 var usuarios = _usuarioRepository.GetUsuarios();
-                return usuarios;
+                if(usuarios.Count == 0)
+                {
+                    throw new Exception("No existen usuarios");
+                }
+                return Ok(usuarios);
             }
             catch (Exception ex)
             {
@@ -90,6 +116,5 @@ namespace BecarioAPI.Controllers
                 return BadRequest(ex.Message);
             }
         }
-       
     }
 }
